Updated decision tree code with modified dataset
