--- edges.csv
@@ -1,40 +1,0 @@
-u,v,type,weight
-BW01,W01,corridor,
-W01,SPW01,corridor,
-SPW01,GW01,corridor,
-GW01,J01,corridor,
-J01,S01,corridor,
-S01,EF01,corridor,
-EF01,R312,corridor,
-R312,J02,corridor,
-J02,R311,corridor,
-R311,R313,corridor,
-R313,R314,corridor,
-R314,L1,corridor,
-L1,L2,corridor,
-L2,L3,corridor,
-L3,S02,corridor,
-S02,UR01,corridor,
-UR01,R315,corridor,
-R315,R316,corridor,
-R316,R310,corridor,
-R310,R309,corridor,
-R309,J03,corridor,
-J03,R317,corridor,
-R317,R318,corridor,
-R318,J04,corridor,
-J04,R319,corridor,
-R319,S03,corridor,
-S03,R301,corridor,
-R301,R302,corridor,
-R302,R303,corridor,
-R303,R304,corridor,
-R304,R308,corridor,
-R308,UR02,corridor,
-UR02,BW02,corridor,
-BW02,R307,corridor,
-R307,W02,corridor,
-W02,GW02,corridor,
-GW02,L4,corridor,
-L4,R305,corridor,
-R305,R306,corridor,
