--- nodes.csv
@@ -1,41 +1,0 @@
-id,name,category,lat,lon,floor
-BW01,Boy's washroom01,washroom,28.687933,77.214142,3
-W01,Water cooler01,Water cooler,28.687902,77.214149,3
-SPW01,Specially able washroom01, washroom,28.688001,77.214099,3
-GW01,girl's washroom01,washroom,28.687944,77.214251,3
-J01,Janitor room01,room,28.687921,77.214152,3
-S01,stairs01,stairs,28.687917,77.214162,3
-EF01,Fire01,fire,28.687898,77.21417,3
-R312,room312,room,28.687922,77.214138,3
-J02,janitor room02,room,28.687966,77.214063,3
-R311,room311,room,28.687884,77.214186,3
-R313,room313,room,28.687868,77.214177,3
-R314,room314,room,28.68789,77.214221,3
-L01,Lift 1,lift,28.687881,77.214198,3
-L02,Lift 2,lift,28.687834,77.214184,3
-L03,Lift 3,lift ,28.687783,77.214135,3
-S02,stairs02,stairs,28.687869,77.214211,3
-UR01,unknown room01,room,28.687783,77.214218,3
-R315,room315,room,28.687744,77.214187,3
-R316,room316,room,28.687735,77.214242,3
-R310,room310,room,28.687751,77.214192,3
-R309,room309,room,28.687725,77.214233,3
-J03,janitor room03,room,28.687735,77.214176,3
-R317,room317,room,28.687713,77.214208,3
-R318,room318,room,28.687675,77.214195,3
-J04,janitor room04,room,28.68767,77.214192,3
-R319,R319,room,28.687664,77.214184,3
-S03,stairs,stairs,28.687646,77.214155,3
-R301,room301,room,28.687585,77.214073,3
-R302,room302,room,28.687561,77.214045,3
-R303,room303,room,28.687573,77.214057,3
-R304,room304,room,28.687493,77.214062,3
-R308,room308,room,28.687467,77.214063,3
-UR02,unknown room02,room,28.687178,77.21405,3
-BW02,Boy's washroom002,washroom,28.687282,77.213955,3
-R307,room307,room,28.687158,77.213972,3
-W02,water cooler02,Water cooler,28.687236,77.213978,3
-GW02,girl's washroom02,washroom,28.687389,77.21408,3
-L4,Lift 4,lift,28.68716,77.213915,3
-R305,room305,room,28.68721,77.21401,3
-R306,room306,room,28.687166,77.213821,3
